closer, but pathing is still wrong
